// web/src/api/generated/model/adminMediaGetAllParams.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * StigVidd
 * Stigvidd API. This is free software, licensed under the GNU Affero General Public License version 3 or later. Under section 13 of that licence you are entitled to the Corresponding Source of this running service: https://github.com/themalind/stigvidd
 * OpenAPI spec version: 1.0.0
 */

export type AdminMediaGetAllParams = {
Page?: number;
PageSize?: number;
/**
 * @nullable
 */
Sort?: string | null;
/**
 * @nullable
 */
OwnerType?: string | null;
/**
 * @nullable
 */
OwnerIdentifier?: string | null;
/**
 * @nullable
 */
Format?: string | null;
/**
 * @nullable
 */
MinWidth?: number | null;
/**
 * @nullable
 */
MaxWidth?: number | null;
/**
 * @nullable
 */
MinHeight?: number | null;
/**
 * @nullable
 */
MaxHeight?: number | null;
/**
 * @nullable
 */
MinSizeBytes?: number | null;
/**
 * @nullable
 */
MaxSizeBytes?: number | null;
/**
 * @nullable
 */
CreatedFrom?: string | null;
/**
 * @nullable
 */
CreatedTo?: string | null;
/**
 * @nullable
 */
TargetMaxWidth?: number | null;
/**
 * @nullable
 */
TargetFormat?: string | null;
};