// web/src/api/generated/model/mediaFilter.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * StigVidd
 * Stigvidd API. This is free software, licensed under the GNU Affero General Public License version 3 or later. Under section 13 of that licence you are entitled to the Corresponding Source of this running service: https://github.com/themalind/stigvidd
 * OpenAPI spec version: 1.0.0
 */

export interface MediaFilter {
  /** @nullable */
  ownerType?: string | null;
  /** @nullable */
  ownerIdentifier?: string | null;
  /** @nullable */
  format?: string | null;
  /** @nullable */
  minWidth?: number | null;
  /** @nullable */
  maxWidth?: number | null;
  /** @nullable */
  minHeight?: number | null;
  /** @nullable */
  maxHeight?: number | null;
  /** @nullable */
  minSizeBytes?: number | null;
  /** @nullable */
  maxSizeBytes?: number | null;
  /** @nullable */
  createdFrom?: string | null;
  /** @nullable */
  createdTo?: string | null;
  /** @nullable */
  targetMaxWidth?: number | null;
  /** @nullable */
  targetFormat?: string | null;
}